Script: new "common.py" module for shared helpers split from "pretext.py"
